requirements.txt
pyotp==2.3.0
flask==2.3.2
flask-Login==0.6.2
gspread==5.9.0
loguru==0.6.0
gunicorn==20.1.0
functions-framework==3.*
google-cloud-storage==2.5.0
google-cloud-bigquery==3.4.1
flask-paginate==2022.1.8
google-api-core==2.11.0
google-api-python-client==1.7.11
google-auth==2.15.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.8.0
Jinja2==3.1.2
google-cloud-logging==3.5.0
Werkzeug==2.3.4
pendulum==2.1.2
